+++ package.json
@@ -3,16 +3,16 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@react-oauth/google": "^0.10.0",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "axios": "^1.3.5",
     "bootstrap": "^5.2.3",
-
-    "pagination": "^0.4.6",
-
     "bootstrap-icons": "^1.10.4",
-
+    "gapi-script": "^1.2.0",
+    "jwt-decode": "^3.1.2",
+    "pagination": "^0.4.6",
     "react": "^18.2.0",
     "react-bootstrap": "^2.7.2",
     "react-bootstrap-icons": "^1.10.3",
