+++ src/App.js
@@ -1,9 +1,9 @@
 import React, { useState, useEffect } from "react";
 import { BrowserRouter, Routes, Route, Link } from "react-router-dom";
 
-import Login from "./components/login";
+import Login from "./components/login/login";
 import Main from "./components/main";
-import KakaoHandler from "./components/social/kakaoHandler";
+import SocialHandler from "./components/social/socialHandler";
 
 import AccountInfo from "./components/accountInfo";
 import MessageInfo from "./components/messageInfo";
@@ -14,7 +14,7 @@ import MyfreebbsList from "./components/myfreebbsList";
 import PartyAccept from "./components/partyAccept";
 import PartyRoom from "./components/partyRoom";
 
-import Regi from "./components/regi";
+import Regi from "./components/login/regi";
 
 import FreeBbsList from "./components/freeBbsList";
 import FreeBbsDetail from "./components/freeBbsDetail";
@@ -84,12 +84,12 @@ function App() {
             <Route exact path="/" element={<Main />} />
             <Route path="/login" element={<Login />} />
             <Route path="/regi" element={<Regi />} />
-            // QNA
+            
             <Route path="/qnalist" element={<Qnalist />} />
             <Route path="/qnawrite" element={<Qnawrite />} />
             <Route path="/qnadetail/:qnaSeq" exact element={<Qnadetail />} />
-            //
-            <Route path="/kakaoLogin" element={<KakaoHandler />} />
+            
+            <Route path="/socialLogin" element={<SocialHandler />} />
             <Route path="/accountInfo" element={<AccountInfo />}></Route>
             <Route path="/mybbsList" element={<MybbsList />}></Route>
             <Route path="/partyAccept" element={<PartyAccept />}></Route>
@@ -100,9 +100,6 @@ function App() {
               element={<SendMessageInfo />}
             ></Route>
             <Route path="/myfreebbsList" element={<MyfreebbsList />}></Route>
-            <Route path="/partyAccept" element={<PartyAccept />}></Route>
-            <Route path="/partyRoom" element={<PartyRoom />}></Route>
-            <Route path="regi" element={<Regi />} />
             <Route path="freeBoard" element={<FreeBbsList />} />
             <Route path="/freeBbsDetail/:bbsSeq" element={<FreeBbsDetail />} />
             <Route path="/freeBbsWrite" element={<FreeBbsWrite />} />
