+++ src/components/login/login.css
@@ -1,0 +1,5 @@
+.google-box{
+    width: 200px;
+    height: 40px;
+    margin: auto;
+}
