+++ src/components/login/regi.js
@@ -2,9 +2,14 @@ import React, { useState, useRef, useCallback } from "react";
 import { useNavigate } from "react-router-dom"
 import axios from "axios";
 import { useDaumPostcodePopup } from 'react-daum-postcode';
+import KakaoLogin from "react-kakao-login";
+// npm install react-kakao-login
+import { GoogleLogin, GoogleOAuthProvider } from '@react-oauth/google';
+import jwt_decode from 'jwt-decode';
+import "./regi.css";
 
 function Regi(){
-    let history = useNavigate();
+    const navigate = useNavigate();
 
     // 각각의 이름 설정
     const [id, setId] = useState('');
@@ -127,7 +132,7 @@ function Regi(){
             .then(function(resp){
                 if(resp.data === "YES"){
                     alert("정상적으로 가입되었습니다.");
-                    history("/");  // 이동(link)
+                    navigate("/");  // 이동(link)
                 }else{
                     alert("가입되지 않았습니다.");
                 }
@@ -200,6 +205,142 @@ function Regi(){
         [password]
     )
 
+    // 카카오 로그인
+    const kakaoClientId = 'bae3cae9292dc7ae67aa897703235e5e'
+    const kakaoOnSuccess = async (data)=>{
+        console.log(data);
+        // alert(data);
+        const idToken = data.response.id_token;  // 인가코드 백엔드로 전달
+        
+        const kakaoId = 'kakao_'+data.profile.id;
+        const kakaoImg = data.profile.properties.profile_image;
+        const kakaoNickname = data.profile.properties.nickname;
+        const kakaoEmail = data.profile.kakao_account.email;
+        // alert(kakaoId);
+        // alert(kakaoNickname);
+
+        console.log(idToken);
+        // alert(idToken);
+
+        // 해당 카카오 계정이 있는지 확인
+        axios.post('http://localhost:3000/idcheck', null, { params:{"id":kakaoId} })
+            .then(function(res){
+                if(res.data === 'YES'){
+                    // 임의 비밀번호
+                    const randomString = Math.random().toString(36).slice(2);
+                
+                    // 데이터를 모아서 백엔드로 넘기기
+                    let kakaoParams = { 'id':kakaoId, 'password':randomString, 'profile':kakaoImg, 'nickname':kakaoNickname, 'email':kakaoEmail, 'phoneNum':'초기값', 'address':'초기값' };
+                    axios.post("http://localhost:3000/socialRegi", null, { params:kakaoParams })
+                      .then(function (resp) {
+                      if (resp.data === "YES") {
+                        console.log(kakaoParams);
+                        navigate('/socialLogin', {
+                          state: {
+                            id: `${kakaoId}`,
+                            nickname:`${kakaoNickname}`
+                          }
+                        });
+                        // document.location.href = `/kakaoLogin`;
+
+                      } else {
+                        alert("가입되지 않았습니다.");
+                      }
+                    })
+                    .catch(function (err) {
+                      alert(err);
+                    });
+                }else{
+                  axios.post("http://localhost:3000/socialLogin", null, { params:{ 'id':kakaoId } })
+                      .then(function (resp) {
+                      console.log(resp.data);
+                      if (resp.data !== null && resp.data !== "") {
+                        alert(resp.data.nickname + "님 환영합니다");
+                        localStorage.setItem("login", JSON.stringify(resp.data));
+            
+                        document.location.href = '/';
+                      } else {
+                        alert("로그인에 실패했습니다.");
+                      }
+                    })
+                    .catch(function (err) {
+                      alert(err);
+                    });
+                  }
+              })
+              .catch(function(err){
+                  alert(err);
+                  alert('아이디 찾기');
+              })
+    }
+    const kakaoOnFailure = (error) => {
+      console.log(error);
+    };
+
+    // 구글 로그인
+    const clientId = '35821294018-ps4nug5312m228unr95fih9gc2hq518u.apps.googleusercontent.com'
+    const googleOnSuccess = async(credentialResponse) => {
+        console.log(credentialResponse.credential);
+        const decodeding = jwt_decode(credentialResponse.credential);
+        console.log(decodeding);
+        
+        const googleId = 'google_'+decodeding.sub;
+        const googleNickname = decodeding.name;
+        const googleEmail = decodeding.email;
+        const googleImg = decodeding.picture;
+
+        // 구글 계정이 있는지 확인
+        axios.post('http://localhost:3000/idcheck', null, { params:{"id":googleId} })
+          .then(function(res){
+              if(res.data === 'YES'){
+                  // 임의 비밀번호
+                  const randomString = Math.random().toString(36).slice(2);
+              
+                  // 데이터를 모아서 백엔드로 넘기기
+                  let googleParams = { 'id':googleId, 'password':randomString, 'profile':googleImg, 'nickname':googleNickname, 'email':googleEmail, 'phoneNum':'초기값', 'address':'초기값' };
+                  axios.post("http://localhost:3000/socialRegi", null, { params:googleParams })
+                    .then(function (resp) {
+                    if (resp.data === "YES") {
+                      console.log(googleParams);
+                      navigate('/socialLogin', {
+                        state: {
+                          id: `${googleId}`,
+                          nickname:`${googleNickname}`
+                        }
+                      });
+                      // document.location.href = `/kakaoLogin`;
+
+                    } else {
+                      alert("가입되지 않았습니다.");
+                    }
+                  })
+                  .catch(function (err) {
+                    alert(err);
+                  });
+              }else{
+                axios.post("http://localhost:3000/socialLogin", null, { params:{ 'id':googleId } })
+                    .then(function (resp) {
+                    console.log(resp.data);
+                    if (resp.data !== null && resp.data !== "") {
+                      alert(resp.data.nickname + "님 환영합니다");
+                      localStorage.setItem("login", JSON.stringify(resp.data));
+          
+                      document.location.href = '/';
+                    } else {
+                      alert("로그인에 실패했습니다.");
+                    }
+                  })
+                  .catch(function (err) {
+                    alert(err);
+                  });
+              }
+          })
+          .catch(function(err){
+              alert(err);
+              alert('아이디 찾기');
+          })
+    }
+
     return (
         <div>
             <h3>회원가입</h3>
@@ -245,6 +386,23 @@ function Regi(){
                 
                 <button type="submit" onClick={account}>회원가입</button>
             </form>
+            <hr/>
+
+            <KakaoLogin
+              token={kakaoClientId}
+              onSuccess={kakaoOnSuccess}
+              onFail={kakaoOnFailure}
+            >카카오로 시작하기</KakaoLogin><br/><br/>
+            <div className="google-box">
+            <GoogleOAuthProvider clientId={clientId}>
+              <GoogleLogin
+                  onSuccess={googleOnSuccess}
+                  onFailure={() => {
+                  console.log('로그인 실패');
+                  }}
+              />
+            </GoogleOAuthProvider>
+            </div>
         </div>
     )
 }
